+++ src/php/Phel.php
@@ -9,6 +9,7 @@ use Gacela\Framework\Bootstrap\GacelaConfig;
 use Gacela\Framework\Gacela;
 use Phar;
 use Phel\Filesystem\FilesystemFacade;
+use Phel\Lang\Registry;
 use Phel\Run\RunFacade;
 
 use function dirname;
@@ -16,7 +17,10 @@ use function in_array;
 use function is_array;
 use function is_string;
 
-final class Phel
+/**
+ * @interal use \Phel instead
+ */
+class Phel
 {
     public const string PHEL_CONFIG_FILE_NAME = 'phel-config.php';
 
@@ -63,6 +67,11 @@ final class Phel
         $runFacade->runNamespace($namespace);
 
         Gacela::get(FilesystemFacade::class)?->clearAll();
+    }
+
+    public static function getDefinition(string $ns, string $name): mixed
+    {
+        return Registry::getInstance()->getDefinition($ns, $name);
     }
 
     /**