+++ tests/php/Integration/Command/Domain/DirectoryFinder/ComposerVendorDirectoriesFinderTest.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace PhelTest\Integration\Command\Domain\DirectoryFinder;
 
 use Gacela\Framework\Gacela;
+use Phel;
 use Phel\Command\Infrastructure\ComposerVendorDirectoriesFinder;
-use Phel\Phel;
 use PHPUnit\Framework\TestCase;
 
 final class ComposerVendorDirectoriesFinderTest extends TestCase